+++ src/h/dir.h
@@ -45,7 +45,7 @@ extern str world_dir_structure[3][NAME_MAX];
 // ---- signatures -------------------------------------------------------------
 void init_paths();
 void init_instance_directory(str *instance_name, u16 *state, u8 STATE_ACTIVE);
-void init_world_directory();
+void init_world_directory(str *world_name);
 
 #define MC_C_DIR_H
 #endif